added permssions using spartie

+++ app/Http/Controllers/ClientController.php
@@ -45,7 +45,7 @@ class ClientController extends Controller
 
     public function destroy(Client $client): RedirectResponse
     {
-        // Gate::authorize(PermissionEnum::DELETE_CLIENTS->value);
+        Gate::authorize(PermissionEnum::DELETE_CLIENTS->value);
 
         $client->delete();
